tomara que ultimas alteracoes

+++ src/main/kotlin/br/com/caethas/atto/atto/modules/caixa/dto/CaixaDto.kt
@@ -2,11 +2,12 @@ package br.com.caethas.atto.atto.modules.caixa.dto
 
 import br.com.caethas.atto.atto.modules.caixa.submodules.transacao.entity.TransacaoEntity
 import br.com.caethas.atto.atto.shared.base.BaseDto
+import java.math.BigDecimal
 import java.time.LocalDateTime
 import java.util.UUID
 
 class CaixaFromDto(
-    val saldoTotal: Double,
+    val saldoTotal: BigDecimal,
 ) : BaseDto.FromDto()
 
 class CaixaToDto(
@@ -15,7 +16,7 @@ class CaixaToDto(
     override val atualizadoEm: LocalDateTime?,
     override val deletadoEm: LocalDateTime?,
     override val sincronizadoEm: LocalDateTime?,
-    val saldoTotal: Double,
+    val saldoTotal: BigDecimal,
 ) : BaseDto.ToDto(
     id,
     criadoEm,
